+++ lib/serega/validations/check_initiate_params.rb
@@ -18,7 +18,7 @@ class Serega
         private
 
         def check_allowed_keys
-          Utils::CheckAllowedKeys.call(opts, serializer_class.config[:initiate_keys])
+          Utils::CheckAllowedKeys.call(opts, serializer_class.config.initiate_keys)
         end
 
         def check_modifiers